handling some exceptions and applying Json/API rules

+++ app/Http/Controllers/CommentController.php
@@ -5,11 +5,16 @@ namespace App\Http\Controllers;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 use App\Http\Controllers\BaseController;
+use Illuminate\Support\Facades\DB;
 
 class CommentController extends BaseController
 {
     public function store(Request $request)
     {
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
         $request->validate([
             'content' => 'required',
             'post_id' => 'required',
@@ -20,61 +25,87 @@ class CommentController extends BaseController
         $comment = Comment::create($comment);
         
         if(is_null($comment))
-            return $this->sendError("Something went wrong while creating");
-
+        return $this->sendError(["Something went wrong while creating"]);
+        
+        $comment['type'] = 'comments';
         return $this->sendResponse($comment,201);
     }
 
-    public function showByPostId($id)
+    public function showByPostId(Request $request, $id)
     {
-        $comments = Comment::where('post_id',$id);
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
+        $comments = Comment::where('post_id',$id)
+        ->select('id','content','is_published','user_id','post_id', DB::raw("'comments' AS type"));
         if(!auth('sanctum')->check())
             $comments = $comments->where("is_published",true);
 
         return $this->sendResponse($comments->get());
     }
 
-    public function showByUserId($id)
+    public function showByUserId(Request $request, $id)
     {
-        $comments = Comment::where('user_id',$id);
+        $requestHeaders = $this->validateHeaders($request);
+
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
+        $comments = Comment::where('user_id',$id)
+        ->select('id','content','is_published','user_id','post_id', DB::raw("'comments' AS type"));
         if(!auth('sanctum')->check())
             $comments = $comments->where("is_published",true);
         
-        return $this->sendResponse($comments->get());
+        $comments = $comments->get();
+        return $this->sendResponse($comments);
     }
 
     public function update(Request $request, $id)
     {
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
         $request->validate([
             'title' => 'required',
             'slug' => 'required',
             'content' => 'required',
             'is_published' => 'required'
         ]);
+
         $comment = Post::find($id);
 
         if($this->isCurrentUserOwner($comment->user_id))
         {
             $is_updated = Comment::find($id)->update($request->all());
             if($is_updated)
-                return $this->sendResponse("Update successfully");
+            {
+                $comment = Comment::find($id);
+                $comment['type'] = 'comments';
+                return $this->sendResponse($comment);
+            }
             
-            return $this->sendError("Something went wrong while updating");
+            return $this->sendError(["Something went wrong while updating"]);
         }
-        return $this->sendError("This comment doesn't belong to you",401);
+        return $this->sendError(["This comment doesn't belong to you"],401);
     }
 
     public function destroy($id)
     {
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
         $comment = Comment::find($id);
         if($this->isCurrentUserOwner($comment->user_id))
         {
             $is_deleted = Comment::find($id)->delete();
             if($is_deleted)
-                return $this->sendResponse("Deleted successfully");
+                return $this->sendResponse(null);
 
-            return $this->sendError("Something went wrong while deleting");
+            return $this->sendError(["Something went wrong while deleting"]);
         }
-        return $this->sendError("This comment doesn't belong to you",401);
+        return $this->sendError(["This comment doesn't belong to you"],401);
     }
 }
