+++ app/Http/Controllers/PostController.php
@@ -7,19 +7,33 @@ use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\BaseController;
+use Illuminate\Support\Facades\DB;
 
 class PostController extends BaseController
 {
-    public function index()
+    
+    public function index(Request $request)
     {
-        $posts = Post::with('latestComments')->withCount('comments');
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
+        $pages = 5;
+        $posts = Post::select("id","title","slug","is_published","content","user_id",DB::raw("'posts' AS type"))
+                ->with('latestComments')
+                ->withCount('comments');
+
         if(!auth('sanctum')->check())
             $posts = $posts->where("is_published",true);
-        return $this->sendResponse($posts->paginate(5));
+        return $this->sendResponse($posts->paginate($pages));
     }
 
     public function store(Request $request)
     {
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
         $request->validate([
             'title' => 'required',
             'slug' => 'required',
@@ -29,17 +43,21 @@ class PostController extends BaseController
         $post = $request->all();
         $post['user_id'] = auth('sanctum')->id();
         if(Post::find(['slug',$post->slug])->exists())
-            return $this->sendError("This record already exists");
+            return $this->sendError(["This record already exists"]);
         $post = Post::create($post);
-
+        
         if(is_null($post))
-            return $this->sendError("Something went wrong while creating");
+            return $this->sendError(["Something went wrong while creating"]);
 
         return $this->sendResponse($post,201);
     }
 
     public function show($id)
     {
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
         $posts = Post::with('comments')->find($id);
         if(!auth('sanctum')->check())
             $posts = $posts->where("is_published",true);
@@ -49,6 +67,10 @@ class PostController extends BaseController
 
     public function update(Request $request, $id)
     {
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
         $request->validate([
             'title' => 'required',
             'slug' => 'required',
@@ -61,23 +83,27 @@ class PostController extends BaseController
         {
             $is_updated = Post::find($id)->update($request->all());
             if($is_updated)
-                return $this->sendResponse("Update successfully");
+                return $this->sendResponse(Post::find($id));
 
-            return $this->sendError("Something went wrong while updating");
+            return $this->sendError(["Something went wrong while updating"]);
         }
-        return $this->sendError("This comment doesn't belong to you",401);
+        return $this->sendError(["This comment doesn't belong to you"],401);
     }
 
     public function destroy($id)
     {
+        $requestHeaders = $this->validateHeaders($request);
+        if(!$requestHeaders['isValid'])
+            return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+
         $post = Post::find($id);
         if(auth('sanctum')->id() === $post->user_id)
         {
             $is_deleted = Post::find($id)->delete();
             if($is_deleted)
-                return $this->sendResponse("Deleted successfully");
-            return $this->sendError("Something went wrong while deleting");
+                return $this->sendResponse(null);
+            return $this->sendError(["Something went wrong while deleting"]);
         }
-        return $this->sendError("This comment doesn't belong to you",401);
+        return $this->sendError(["This comment doesn't belong to you"],401);
     }
 }
